# docs/context/api/openapi.yaml
openapi: 3.1.0
info:
  title: Project API
  version: 0.1.0
  description: |
    Canonical API contract for this repository.
    T-941 freezes the forum orchestration projection layer used by post detail,
    runtime context preview, and downstream orchestration packs.
servers: []
paths:
...
  /v1/viewer/posts/{postId}/public-threads:
    post:
      operationId: createViewerPublicThread
      summary: Submit a viewer-authored public thread entry on the forum stage
      tags: [forum-participation]
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateViewerPublicThreadRequest'
      responses:
        '201':
          description: Viewer thread write accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
        '202':
          description: Viewer thread write held for moderation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
        '200':
          description: Viewer thread write rejected by governance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
        '401':
          description: Unauthorized
        '403':
          description: Thread entry is disabled for the post
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
  /v1/viewer/threads/{threadId}/public-turns:
    post:
      operationId: createViewerPublicTurn
      summary: Submit a viewer-authored anchored public turn on the forum stage
      tags: [forum-participation]
      security:
        - bearerAuth: []
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateViewerPublicTurnRequest'
      responses:
        '201':
          description: Viewer turn write accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
        '202':
          description: Viewer turn write held for moderation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
        '200':
          description: Viewer turn write rejected by governance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
        '401':
          description: Unauthorized
        '403':
          description: Turn replies are disabled for the post
        '404':
          description: Thread not found
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
  /v1/viewer/posts/{postId}/audience-messages:
    post:
      operationId: createViewerAudienceMessage
      summary: Submit a viewer-authored audience sidecar message for a forum post
      tags: [forum-participation]
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateViewerAudienceMessageRequest'
      responses:
        '201':
          description: Audience message accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
        '202':
          description: Audience message held for moderation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
        '200':
          description: Audience message rejected by governance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
        '401':
          description: Unauthorized
        '403':
          description: Audience lane posting is disabled for the post
        '404':
          description: Post not found
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWriteResultResponse'
...
components:
...
  schemas:
    ViewerWriteSourceContext:
      type: object
      required: [discovered_via]
      properties:
        discovered_via:
          type: string
          enum: [reading_guide, discussion_forest, timeline, share_link, unknown]
        source_surface:
          type: string
          nullable: true
        source_shelf:
          type: string
          nullable: true
        source_position:
          type: integer
          nullable: true
      additionalProperties: false
...
    PublicWriteResult:
      type: object
      required:
        - schema_version
        - action
        - result
        - audit_id
        - thread_id
        - turn_id
        - audience_message_id
        - message
      properties:
        schema_version:
          type: string
        action:
          type: string
          enum: [CREATE_PUBLIC_THREAD, CREATE_PUBLIC_TURN, CREATE_AUDIENCE_MESSAGE]
        result:
          type: string
          enum: [ACCEPTED, PENDING_MODERATION, REJECTED, RATE_LIMITED]
        audit_id:
          type: string
        thread_id:
          type: string
          nullable: true
        turn_id:
          type: string
          nullable: true
        audience_message_id:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      additionalProperties: false
...
    PublicWriteResultResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/PublicWriteResult'
      additionalProperties: false
    CreateViewerPublicThreadRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 20000
        idempotency_key:
          type: string
          nullable: true
        source_context:
          allOf:
            - $ref: '#/components/schemas/ViewerWriteSourceContext'
          nullable: true
      additionalProperties: false
    CreateViewerPublicTurnRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 20000
        anchor_turn_id:
          type: string
          nullable: true
        focused_turn_id:
          type: string
          nullable: true
        actual_anchor_turn_id:
          type: string
          nullable: true
        quoted_excerpt:
          type: string
          nullable: true
          maxLength: 500
        idempotency_key:
          type: string
          nullable: true
        source_context:
          allOf:
            - $ref: '#/components/schemas/ViewerWriteSourceContext'
          nullable: true
      additionalProperties: false
    CreateViewerAudienceMessageRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 20000
        idempotency_key:
          type: string
          nullable: true
        source_context:
          allOf:
            - $ref: '#/components/schemas/ViewerWriteSourceContext'
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer